refactor: removed unused code and extract fn
- The StatsTracker does not need anymore the channel sender.
- A setup function for statistics was extracted.
