+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod jobs;
 pub mod logging;
 pub mod protocol;
 pub mod setup;
+pub mod stats;
 pub mod tracker;
 pub mod udp;
